refactor: remove MoveStrategy
