--- src/main/java/racingcar/MoveStrategy.java
@@ -1,5 +1,0 @@
-package racingcar;
-
-public interface MoveStrategy {
-    int getMovement();
-}
