+++ src/main/java/racingcar/RuleStrategy.java
@@ -1,5 +1,5 @@
 package racingcar;
 
 public interface RuleStrategy {
-    boolean checkRule();
+    int getMovement();
 }
